dhcp: Allow ':' in hexadecimal DUID (dnsmasq)
dnsmasq needs DUIDs to be in colon ':' delimited
format. The fronted shows DUIDs in the static leases
dropdown with ':' sperators, when selecting it errors
and won't let you edit. This will strip out the ':'
before verifying with the validator.

The related fix for dnsmasq is also submitted for
/etc/init.d/dnsmasq under d6a938eb2bb7591682febfc8539adac0ec76a1d6
It takes the hexadecimal string, and colon delimits it.